feat: enhance comparison functionality and UI, update styles and dialog behavior
