+++ src/app/comparision/components/comparision-preview/comparision-preview.component.css
@@ -86,8 +86,10 @@ h1[mat-dialog-title] {
 .properties-grid {
   display: grid;
   grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 4px;
+  padding: 16px;
+  gap: 16px;
   justify-content: center;
+  overflow-y: auto;
 }
 
 .property-card {
